read null

+++ src/cr_API/cr_API.h
@@ -26,7 +26,7 @@ void cr_bitmap(unsigned disk, bool hex);
 int cr_exists(unsigned disk, char* filename);
 int cr_write(crFILE* file_desc, void* buffer, int nbytes);
 int cr_close(crFILE* file_desc);
-long int get_free_block(unsigned int puntero_indice);
+unsigned int get_free_block(unsigned int puntero_indice);
 
 void cr_ls(unsigned disk);
 crFILE* cr_open(unsigned disk, char* filename, char* mode);
@@ -39,4 +39,3 @@ int cr_softlink(unsigned disk_orig, unsigned disk_dest, char* orig, char* dest);
 
 int cr_unload(unsigned disk, char* orig, char* dest);
 int cr_load(unsigned disk, char* orig);
-
